Added incoming_flows_count to node config. Port config needs to be cleaned.
Introduces an incoming_flows_count field to the node startup configuration. This allows the dataplane to create the correct number of server sockets based on the number of incoming flows for each node.

 From the docs:
"Note that, for listening sockets, there is no 'backlog'; to be able to simultaneously accept several connections, as many sockets must be allocated, or any new connection attempts will be reset."